pyproject.toml
[tool.poetry]
name = "msothonewspy"
version = "1.1.1"
description = "The news client written in Python that fetches and curates the world news across the web."
authors = ["Sechaba Mofokeng <sechaba@onemoola.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/msotho/newspy"
repository = "https://github.com/msotho/newspy"
documentation = "https://github.com/msotho/newspy#readme"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/msotho/newspy/issues"


[tool.poetry.dependencies]
python = "^3.10"
python-slugify = "6.1.2"
requests = "2.28.1"
urllib3 = "1.26.12"
feedparser = "6.0.10"

[tool.poetry.group.dev.dependencies]
black = "22.10.0"
pre-commit = "2.20.0"
pytest = "7.2.0"
responses = "0.22.0"
pytest-cov = "4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
